+++ tests/inkscape/test_runner.py
@@ -157,7 +157,8 @@ class TestInkscape(object):
         inkscape.export_pdf(tree, output_path)
         assert len(runner.commands) == 1
         args = runner.commands[0]
-        assert args[1] == '--export-pdf=%s' % output_path
+        assert args[1] == '--export-area-page'
+        assert args[2] == '--export-pdf=%s' % output_path
         assert tmpdir.join('foo').isdir()
 
     def test_contextmanager(self, inkscape, runner):